Create solution for 03-flex-header-2
Adjusted and added CSS rules in order to reach the desired outcome.
The solution was not viewed until completed, explaining the differences in code.
One might argue that because my code uses a margin with auto that it is not completely flexbox. However, I believe due to previous examples used, that this is perfectly acceptable for flexbox to push a group of elements to an edge. But, it is also understood the grouping and using space-between is more flexbox-like than my solution and can provide more segregation between the groups of elements, arguing it would be more human-readible.
That said, since I looked at the solution, I won't be retracting it, but it's important to understand the differences, which is accomplished.