Create solution for 06-flex-layout
Adjusted only CSS to make the content match the desired-outcome. The solution was not viewed before covering the self-check. My solution came out pretty close to what was in the actual solution. Major diffrence being I used jusify-content to force the header and footers to the ends, while the actual solution uses flex on the content to push them away.